swarm/storage/script: Added bzz-script to API constructor
